Changes per trello

+++ database/migrations/2015_12_06_180607_create_products_table.php
@@ -16,7 +16,7 @@ class CreateProductsTable extends Migration
             $table->increments('id');
             $table->timestamps();
             $table->text('name');
-            $table->integer('price');
+            $table->decimal('price', 5,2);
 
         });
     }
